+++ Parchment/Classes/PagingCollectionViewLayout.swift
@@ -21,6 +21,7 @@ open class PagingCollectionViewLayout<T: PagingItem>:
   var state: PagingState<T>?
   var dataStructure: PagingDataStructure<T>
   var layoutAttributes: [IndexPath: PagingCellLayoutAttributes] = [:]
+  var currentTransition: PagingTransition? = nil
   
   open override var collectionViewContentSize: CGSize {
     return contentSize
@@ -60,7 +61,7 @@ open class PagingCollectionViewLayout<T: PagingItem>:
   private let borderLayoutAttributes: PagingBorderLayoutAttributes
   private var contentSize: CGSize = .zero
   private var invalidationSummary: InvalidationSummary = .everything
-  private var currentTransition: PagingTransition? = nil
+  
   private let PagingIndicatorKind = "PagingIndicatorKind"
   private let PagingBorderKind = "PagingBorderKind"
 
@@ -98,18 +99,9 @@ open class PagingCollectionViewLayout<T: PagingItem>:
     super.prepare()
     
     switch invalidationSummary {
-    case .everything, .dataSourceCounts:
+    case .everything, .dataSourceCounts, .sizes:
       layoutAttributes = [:]
       createLayoutAttributes()
-    case .contentOffset:
-      invalidateContentOffset()
-    case .transition:
-      invalidateTransition()
-      invalidateContentOffset()
-    case .sizes:
-      layoutAttributes = [:]
-      createLayoutAttributes()
-      invalidateContentOffset()
     case .partial:
       break
     }
@@ -169,208 +161,8 @@ open class PagingCollectionViewLayout<T: PagingItem>:
     
     return layoutAttributes
   }
-  
-  // The content offset and distance between items can change while a
-  // transition is in progress meaning the current transition will be
-  // wrong. For instance, when hitting the edge of the collection view
-  // while transitioning we need to reload all the paging items and
-  // update the transition data.
-  func updateCurrentTransition() {
-    let oldTransition = currentTransition
-    invalidateTransition()
-    
-    if let oldTransition = oldTransition,
-      let newTransition = currentTransition {
-      
-      let contentOffset = CGPoint(
-        x: view.contentOffset.x - (oldTransition.distance - newTransition.distance),
-        y: view.contentOffset.y)
-      
-      currentTransition = PagingTransition(
-        contentOffset: contentOffset,
-        distance: oldTransition.distance)
-    }
-  }
 
   // MARK: Private
-  
-  private func invalidateContentOffset() {
-    guard let state = state else { return }
-    
-    if options.menuTransition == .scrollAlongside {
-      if let transition = currentTransition {
-        
-        // FIXME: Remove the progress check and move state into
-        // invalidation context.
-        if contentSize.width >= view.bounds.width && state.progress != 0 {
-          let contentOffset = CGPoint(
-            x: transition.contentOffset.x + (transition.distance * fabs(state.progress)),
-            y: transition.contentOffset.y)
-          
-          // We need to use setContentOffset with no animation in
-          // order to stop any ongoing scroll.
-          view.setContentOffset(contentOffset, animated: false)
-        }
-      }
-    }
-  }
-  
-  /// In order to get the menu items to scroll alongside the content
-  /// we create a transition struct to keep track of the initial
-  /// content offset and the distance to the upcoming item so that we
-  /// can update the content offset as the user is swiping.
-  private func invalidateTransition() {
-    guard
-      let state = state,
-      let upcomingPagingItem = state.upcomingPagingItem,
-      let upcomingIndexPath = dataStructure.indexPathForPagingItem(upcomingPagingItem),
-      let to = layoutAttributes[upcomingIndexPath] else {
-        
-        // When there is no upcomingIndexPath or any layout attributes
-        // for that item we have no way to determine the distance.
-        currentTransition = PagingTransition(
-          contentOffset: view.contentOffset,
-          distance: 0)
-        return
-    }
-    
-    var distance: CGFloat = 0
-    
-    switch (options.selectedScrollPosition) {
-    case .left:
-      distance = distanceToLeftAlignedItem()
-    case .right:
-      distance = distanceToRightAlignedItem()
-    case .preferCentered:
-      distance = distanceToCenteredItem()
-    }
-    
-    // Update the distance to account for cases where the user has
-    // scrolled all the way over to the other edge.
-    if view.near(edge: .left, clearance: -distance) && distance < 0 && dataStructure.hasItemsBefore == false {
-      distance = -(view.contentOffset.x + view.contentInset.left)
-    } else if view.near(edge: .right, clearance: distance) && distance > 0 &&
-      dataStructure.hasItemsAfter == false {
-      
-      distance = view.contentSize.width - (view.contentOffset.x + view.bounds.width)
-      
-      if sizeCache.implementsWidthDelegate {
-        let toWidth = sizeCache.itemWidthSelected(for: upcomingPagingItem)
-        distance += toWidth - to.frame.width
-        
-        if let currentIndexPath = dataStructure.indexPathForPagingItem(state.currentPagingItem),
-          let from = layoutAttributes[currentIndexPath] {
-          let fromWidth = sizeCache.itemWidth(for: state.currentPagingItem)
-          distance -= from.frame.width - fromWidth
-        }
-        
-        // If the selected cells grows so much that it will move
-        // beyond the center of the view, we want to update the
-        // distance after all.
-        if options.selectedScrollPosition == .preferCentered {
-          let center = view.bounds.midX
-          let centerAfterTransition = to.frame.midX - distance
-          if centerAfterTransition < center {
-            distance = view.contentSize.width - (view.contentOffset.x + view.bounds.width)
-          }
-        }
-      }
-    }
-    
-    currentTransition = PagingTransition(
-      contentOffset: view.contentOffset,
-      distance: distance)
-  }
-  
-  private func distanceToLeftAlignedItem() -> CGFloat {
-    guard
-      let state = state,
-      let upcomingPagingItem = state.upcomingPagingItem,
-      let upcomingIndexPath = dataStructure.indexPathForPagingItem(upcomingPagingItem),
-      let to = layoutAttributes[upcomingIndexPath] else { return 0 }
-    
-    var distance = to.frame.origin.x - view.contentOffset.x
-    
-    if sizeCache.implementsWidthDelegate {
-      if let currentIndexPath = dataStructure.indexPathForPagingItem(state.currentPagingItem),
-        let from = layoutAttributes[currentIndexPath] {
-        if upcomingPagingItem > state.currentPagingItem {
-          let fromWidth = sizeCache.itemWidth(for: state.currentPagingItem)
-          let fromDiff = from.frame.width - fromWidth
-          distance -= fromDiff
-        }
-      }
-    }
-    return distance
-  }
-  
-  private func distanceToRightAlignedItem() -> CGFloat {
-    guard
-      let state = state,
-      let upcomingPagingItem = state.upcomingPagingItem,
-      let upcomingIndexPath = dataStructure.indexPathForPagingItem(upcomingPagingItem),
-      let to = layoutAttributes[upcomingIndexPath] else { return 0 }
-    
-    let toWidth = sizeCache.itemWidthSelected(for: upcomingPagingItem)
-    let currentPosition = to.frame.origin.x + to.frame.width
-    let width = view.contentOffset.x + view.bounds.width
-    var distance = currentPosition - width
-    
-    if sizeCache.implementsWidthDelegate {
-      if let currentIndexPath = dataStructure.indexPathForPagingItem(state.currentPagingItem),
-        let from = layoutAttributes[currentIndexPath] {
-        if upcomingPagingItem < state.currentPagingItem {
-          let toDiff = toWidth - to.frame.width
-          distance += toDiff
-        } else {
-          let fromWidth = sizeCache.itemWidth(for: state.currentPagingItem)
-          let fromDiff = from.frame.width - fromWidth
-          let toDiff = toWidth - to.frame.width
-          distance -= fromDiff
-          distance += toDiff
-        }
-      } else {
-        distance += toWidth - to.frame.width
-      }
-    }
-    
-    return distance
-  }
-  
-  private func distanceToCenteredItem() -> CGFloat {
-    guard
-      let state = state,
-      let upcomingPagingItem = state.upcomingPagingItem,
-      let upcomingIndexPath = dataStructure.indexPathForPagingItem(upcomingPagingItem),
-      let to = layoutAttributes[upcomingIndexPath] else { return 0 }
-    
-    let toWidth = sizeCache.itemWidthSelected(for: upcomingPagingItem)
-    var distance = to.frame.midX - view.bounds.midX
-    
-    if let currentIndexPath = dataStructure.indexPathForPagingItem(state.currentPagingItem),
-      let from = layoutAttributes[currentIndexPath] {
-      
-      let distanceToCenter = view.bounds.midX - from.frame.midX
-      let distanceBetweenCells = to.frame.midX - from.frame.midX
-      distance = distanceBetweenCells - distanceToCenter
-      
-      if sizeCache.implementsWidthDelegate {
-        let fromWidth = sizeCache.itemWidth(for: state.currentPagingItem)
-        
-        if upcomingPagingItem < state.currentPagingItem {
-          distance = -(to.frame.width + (from.frame.midX - to.frame.maxX) - (toWidth / 2)) - distanceToCenter
-        } else {
-          let toDiff = (toWidth - to.frame.width) / 2
-          distance = fromWidth + (to.frame.midX - from.frame.maxX) + toDiff - (from.frame.width / 2) - distanceToCenter
-        }
-      }
-    } else if sizeCache.implementsWidthDelegate {
-      let toDiff = toWidth - to.frame.width
-      distance += toDiff / 2
-    }
-    
-    return distance
-  }
   
   private func createLayoutAttributes() {
     guard let state = state else { return }
