+++ Parchment/Classes/PagingInvalidationContext.swift
@@ -1,7 +1,5 @@
 import UIKit
 
 open class PagingInvalidationContext: UICollectionViewLayoutInvalidationContext {
-  var invalidateContentOffset: Bool = false
-  var invalidateTransition: Bool = false
   var invalidateSizes: Bool = false
 }
